requirements.txt
asyncpg
discord
DateTime
asyncio
statcord.py
topggpy
python-dotenv
pyyaml
coloredlogs